move Schema class from types_.py -> schema.py
